fix test and cleanup (#1)
Fix `allowGlobalAutomatic` mode for interfaces (#466)
Fix JNI caching for native threads (#464)
fix extensions (#473)
fix protocols that return java classes. (#479)

+++ Samples/SwiftJavaExtractJNISampleApp/Sources/MySwiftLibrary/LibrarySubDirectory/SwiftTypeInSubDirectory.swift
@@ -15,7 +15,7 @@
 public final class SwiftTypeInSubDirectory {
   public init() {}
 
-  public func hello() -> Int {
+  public func hello() -> Int64 {
     12
   }
 }
